Fix issue with location being nil on update
A few users have tried clicking 'Next' on the Location screen after
signing in for the first time without entering any location information.
The first time, they get an error about needing to enter a value for at
least one field. If they click Save without entering anything, a 500
error was occurring because the location existed in the session, but it
wasn't associated with the user.